MicroservicesDeployment/frontend/requirements.txt
flask ~= 1.1
grpcio-tools ~= 1.30
jinja2 ~= 2.11
pytest ~= 5.4